allocate home and landlord pages in the pages folder for proper routing
