+++ frontend/src/components/Layout.tsx
@@ -1,19 +1,19 @@
 import { Outlet, Route, Routes } from "react-router"
 import Header from "./Header"
-import HomePage from "./HomePage"
+import HomePage from "../pages/Home"
 
 const Layout = () => {
   return (
     <div>
-        <Header />
-        <div>
-            <Outlet />
-        </div>
-        <Routes>
-            <Route index element={<HomePage />} />
-        </Routes>
+      <Header />
+      <div>
+        <Outlet />
+      </div>
+      <Routes>
+        <Route index element={<HomePage />} />
+      </Routes>
     </div>
-    
+
   )
 }
