+++ frontend/src/pages/Landlord/Landlord.tsx
@@ -1,0 +1,19 @@
+import { Link, Route, Routes } from "react-router";
+import LandlordHeader from "../../components/landlord/LandlordHeader";
+
+
+const LandlordPage = () => {
+  return (
+    <div className="flex justify-center items-center h-screen">
+      <nav>
+        <Link to={'/landlord'}>Dashboard</Link>
+      </nav>
+
+      <Routes>
+        <Route index element={<LandlordHeader />} />
+      </Routes>
+    </div>
+  );
+}
+
+export default LandlordPage
